Add DenseNet models
This adds the densenet121 and densenet169 models to the project. Both models
support pretraining with imagenet weights. The link to these weights can be
found at the top level comment of each model. They must be locally downloaded
and accessible in the docker via the file path `/opt/data/weights/`.